Extending time_t to 64bits.

+++ drivers/vbox/vbox.c
@@ -161,12 +161,12 @@ static void vbox_update_time(void)
 			sizeof(*req)) == VMMDEV_ERR_OK) {
 		time(&otime);				/* old time */
 
-		ntime = (unsigned long)(req->time / 1000);	/* new time */
+		ntime = req->time / 1000;		/* new time */
 
 		/* Make time go forward, if the difference exceeds the drift
 		 * threshold. Never make time go backward.
 		 */
-		if ((int) (ntime - otime) >= drift)
+		if ((ntime - otime) >= drift)
 			stime(&ntime);
 	}
