+++ lib/libc/sys-minix/utimes.c
@@ -31,10 +31,11 @@ int utimes(const char *name, const struct timeval tv[2])
 	m.VFS_UTIMENS_ANSEC = m.VFS_UTIMENS_MNSEC = UTIME_NOW;
   }
   else {
-	m.VFS_UTIMENS_ATIME = tv[0].tv_sec;
-	m.VFS_UTIMENS_MTIME = tv[1].tv_sec;
-	m.VFS_UTIMENS_ANSEC = tv[0].tv_usec * 1000;
-	m.VFS_UTIMENS_MNSEC = tv[1].tv_usec * 1000;
+	/* For now just truncate time_t values to 32bits. */
+	m.VFS_UTIMENS_ATIME = (int32_t)tv[0].tv_sec;
+	m.VFS_UTIMENS_MTIME = (int32_t)tv[1].tv_sec;
+	m.VFS_UTIMENS_ANSEC = (int32_t)tv[0].tv_usec * 1000;
+	m.VFS_UTIMENS_MNSEC = (int32_t)tv[1].tv_usec * 1000;
   }
   m.VFS_UTIMENS_FLAGS = 0;
